edit implementation and move to misc reference
