+++ native/spark-expr/src/lib.rs
@@ -19,7 +19,9 @@
 // The lint makes easier for code reader/reviewer separate references clones from more heavyweight ones
 #![deny(clippy::clone_on_ref_ptr)]
 
+mod downcast;
 mod error;
+pub(crate) use downcast::{downcast_named_arg, opt_downcast_arg};
 mod query_context;
 
 pub mod kernels;
@@ -58,6 +60,7 @@ pub use bloom_filter::{BloomFilterAgg, BloomFilterMightContain};
 mod conditional_funcs;
 mod conversion_funcs;
 mod math_funcs;
+mod misc_funcs;
 mod nondetermenistic_funcs;
 
 pub use array_funcs::*;
@@ -78,12 +81,13 @@ pub use error::{SparkError, SparkErrorWithContext, SparkResult};
 pub use hash_funcs::*;
 pub use json_funcs::{FromJson, ToJson};
 pub use math_funcs::{
-    create_modulo_expr, create_negate_expr, spark_aes_decrypt, spark_ceil, spark_decimal_div,
+    create_modulo_expr, create_negate_expr, spark_ceil, spark_decimal_div,
     spark_decimal_integral_div, spark_floor, spark_make_decimal, spark_round, spark_unhex,
     spark_unscaled_value, CheckOverflow, DecimalRescaleCheckOverflow, NegativeExpr,
     NormalizeNaNAndZero, WideDecimalBinaryExpr, WideDecimalOp,
 };
 pub use query_context::{create_query_context_map, QueryContext, QueryContextMap};
+pub use misc_funcs::spark_aes_decrypt;
 pub use string_funcs::*;
 
 /// Spark supports three evaluation modes when evaluating expressions, which affect
