+++ native/spark-expr/src/misc_funcs/mod.rs
@@ -15,26 +15,6 @@
 // specific language governing permissions and limitations
 // under the License.
 
-pub(crate) mod abs;
-mod ceil;
-pub(crate) mod checked_arithmetic;
-mod div;
-mod floor;
-pub mod internal;
-pub mod modulo_expr;
-mod negative;
-mod round;
-pub(crate) mod unhex;
-mod utils;
-mod wide_decimal_binary_expr;
+pub(crate) mod aes_decrypt;
 
-pub use ceil::spark_ceil;
-pub use div::spark_decimal_div;
-pub use div::spark_decimal_integral_div;
-pub use floor::spark_floor;
-pub use internal::*;
-pub use modulo_expr::create_modulo_expr;
-pub use negative::{create_negate_expr, NegativeExpr};
-pub use round::spark_round;
-pub use unhex::spark_unhex;
-pub use wide_decimal_binary_expr::{WideDecimalBinaryExpr, WideDecimalOp};
+pub use aes_decrypt::spark_aes_decrypt;
